BirthdayCake_Problem
HackerRank keeps saying there is a run time error with my solution, the error occurs on their test cases that contain lists with 100,000+ members. I have tested my solution with lists that contain over 10,000 members and I'm fairly certain that it should work no matter the length of the list.